+++ BasicPythonScripts/RandomNumberGenerator.py
@@ -5,7 +5,7 @@ enternumber = 0
   
 while enternumber == 0:
   try:
-    range_low = int(input("Lowest number of range: "))
+    range_low = float(input("Lowest number of range: "))
     enternumber = 1
   except:
     print("Bummer. That broke. Please enter a numeric value.")
@@ -15,7 +15,7 @@ enternumber = 0
 
 while enternumber == 0:
   try:
-    range_high = int(input("Highest number of range: "))
+    range_high = float(input("Highest number of range: "))
     enternumber = 1
   except:
     print("Bummer. That broke. Please enter a numeric value.")
@@ -25,7 +25,7 @@ enternumber = 0
 
 while enternumber == 0:
   try: 
-    amount_of_variables = int(input("How many numbers do you want? "))
+    amount_of_variables = float(input("How many numbers do you want? "))
     enternumber = 1
   except:
     print("Bummer. That broke. Please enter a numeric value.")
